refactor(models): privatize xr-ai-models implementation modules
Move the xr-ai-models implementation modules to underscore-private names,
matching the render-subagents branch layout:
  config.py       -> _config.py
  factory.py      -> _factory.py
  openai_compat.py-> _openai_compat.py
  protocols.py    -> _protocols.py

The package's public API is unchanged: every name previously re-exported from
`xr_ai_models` is still importable from the package root. In-tree consumers now
import from the package root (vlm-mcp drops the redundant explicit
`kind=KIND_OPENAI_COMPAT`, which is already the VLMSpec default; tests reach
into the private modules for internal helpers).

Because the four old module paths were documented public imports, each is kept
as a deprecated forwarding alias that emits a DeprecationWarning and re-exports
the canonical objects (`KIND_OPENAI_COMPAT`/`ModelKind` stay re-exported from
`xr_ai_models.config`). New test asserts the aliases forward and warn and that
the public names remain reachable from the package root.

The ruff.toml grandfather snapshot entries follow the rename (module contents
unchanged, so their pre-existing I001/E402 ignores move to the new paths).
DEPENDENCIES.md and the xr-ai-models README reference `_factory.py::make_*`.
