+++ agent-sdk/xr-ai-models/xr_ai_models/factory.py
@@ -1,69 +1,24 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""``make_*`` constructors that dispatch a :class:`Spec` to a concrete client."""
+"""Deprecated forwarding alias for the service factory helpers.
+
+The ``make_*`` factory helpers moved to the private :mod:`xr_ai_models._factory`.
+Import them from the package root (``from xr_ai_models import make_vlm``) instead;
+this alias will be removed in a future version.
+"""
+
 from __future__ import annotations
 
-from .config import KIND_OPENAI_COMPAT, ModelsConfig
-from .openai_compat import (
-    OpenAICompatLLM,
-    OpenAICompatSTT,
-    OpenAICompatTTS,
-    OpenAICompatVLM,
+import warnings
+
+from ._factory import make_llm, make_stt, make_tts, make_vlm
+
+warnings.warn(
+    "xr_ai_models.factory is deprecated; import from xr_ai_models instead. "
+    "This alias will be removed in a future version.",
+    DeprecationWarning,
+    stacklevel=2,
 )
-from .protocols import Capabilities, LLMService, STTService, TTSService, VLMService
 
-
-def make_llm(config: ModelsConfig, name: str) -> LLMService:
-    spec = config.llm(name)
-    if spec.kind == KIND_OPENAI_COMPAT:
-        return OpenAICompatLLM(
-            base_url=spec.base_url,
-            model_name=spec.model_name,
-            capabilities=Capabilities(**spec.capabilities),
-            reasoning_field=spec.reasoning_field,
-            default_extras=spec.default_extras,
-            api_key_env=spec.api_key_env,
-            timeout=spec.timeout,
-            health_check=spec.health_check,
-        )
-    raise ValueError(f"unsupported LLM kind: {spec.kind!r}")
-
-
-def make_vlm(config: ModelsConfig, name: str) -> VLMService:
-    spec = config.vlm(name)
-    if spec.kind == KIND_OPENAI_COMPAT:
-        return OpenAICompatVLM(
-            base_url=spec.base_url,
-            model_name=spec.model_name,
-            capabilities=Capabilities(**spec.capabilities),
-            default_extras=spec.default_extras,
-            api_key_env=spec.api_key_env,
-            timeout=spec.timeout,
-            health_check=spec.health_check,
-        )
-    raise ValueError(f"unsupported VLM kind: {spec.kind!r}")
-
-
-def make_stt(config: ModelsConfig, name: str) -> STTService:
-    spec = config.stt(name)
-    if spec.kind == KIND_OPENAI_COMPAT:
-        return OpenAICompatSTT(
-            base_url=spec.base_url,
-            api_key_env=spec.api_key_env,
-            timeout=spec.timeout,
-            health_check=spec.health_check,
-        )
-    raise ValueError(f"unsupported STT kind: {spec.kind!r}")
-
-
-def make_tts(config: ModelsConfig, name: str) -> TTSService:
-    spec = config.tts(name)
-    if spec.kind == KIND_OPENAI_COMPAT:
-        return OpenAICompatTTS(
-            base_url=spec.base_url,
-            api_key_env=spec.api_key_env,
-            timeout=spec.timeout,
-            health_check=spec.health_check,
-        )
-    raise ValueError(f"unsupported TTS kind: {spec.kind!r}")
+__all__ = ["make_llm", "make_stt", "make_tts", "make_vlm"]
