+++ ruff.toml
@@ -42,10 +42,10 @@ known-first-party = ["_helpers", "_helpers_subprocess"]
 "agent-samples/xr-render-demo/worker/tooling.py" = ["E701", "E702", "I001"]
 "agent-samples/xr-render-demo/worker/xr_render_demo_worker.py" = ["I001"]
 "agent-sdk/xr-ai-models/xr_ai_models/__init__.py" = ["I001"]
-"agent-sdk/xr-ai-models/xr_ai_models/config.py" = ["I001"]
-"agent-sdk/xr-ai-models/xr_ai_models/openai_compat.py" = ["E402", "I001"]
+"agent-sdk/xr-ai-models/xr_ai_models/_config.py" = ["I001"]
+"agent-sdk/xr-ai-models/xr_ai_models/_openai_compat.py" = ["E402", "I001"]
+"agent-sdk/xr-ai-models/xr_ai_models/_protocols.py" = ["I001"]
 "agent-sdk/xr-ai-models/xr_ai_models/presets/__init__.py" = ["I001"]
-"agent-sdk/xr-ai-models/xr_ai_models/protocols.py" = ["I001"]
 "agent-sdk/xr-ai-pipecat/xr_ai_pipecat/audio.py" = ["I001"]
 "agent-sdk/xr-ai-pipecat/xr_ai_pipecat/pipeline.py" = ["I001"]
 "agent-sdk/xr-ai-pipecat/xr_ai_pipecat/processors/streaming_tts.py" = ["I001"]
